+++ python/mujoco/introspect/structs.py
@@ -343,6 +343,11 @@ STRUCTS: Mapping[str, StructDecl] = dict([
                  doc='maximum number of convex collision solver iterations',
              ),
              StructFieldDecl(
+                 name='onwarn',
+                 type=ValueType(name='int'),
+                 doc='response to simulation warnings (mjtOnWarn)',
+             ),
+             StructFieldDecl(
                  name='disableflags',
                  type=ValueType(name='int'),
                  doc='bit flags for disabling standard features',
@@ -5727,6 +5732,16 @@ STRUCTS: Mapping[str, StructDecl] = dict([
                      extents=(15,),
                  ),
                  doc='timer statistics',
+             ),
+             StructFieldDecl(
+                 name='status',
+                 type=ValueType(name='int'),
+                 doc='status of the last pipeline call (mjtStatus)',
+             ),
+             StructFieldDecl(
+                 name='nested',
+                 type=ValueType(name='int'),
+                 doc='nonzero while a pipeline call is in progress',
              ),
              StructFieldDecl(
                  name='ncon',