+++ src/engine/engine_core_smooth.c
@@ -1978,6 +1978,7 @@ void mj_makeM(const mjModel* m, mjData* d) {
 // sparse L'*D*L factorizaton of the inertia matrix M, assumed spd
 void mj_factorM(const mjModel* m, mjData* d) {
   TM_START;
+  mjENTER(d);
 
   // sleep filtering
   int sleep_filter = mjENABLED(mjENBL_SLEEP) && d->nv_awake < m->nv;
@@ -2008,6 +2009,7 @@ void mj_factorM(const mjModel* m, mjData* d) {
   }
 
   TM_ADD(mjTIMER_POS_INERTIA);
+  mjLEAVE(d);
 }